+++ LearnBot/src/main/java/org/firstinspires/ftc/learnbot/Robot.java
@@ -41,7 +41,7 @@ public class Robot implements Loggable {
     public void atStart() {}
 
     public void updatePose() {
-        Pose pose = getFollower().getPose();
+        Pose pose = follower.getPose();
         gyro = pose.getHeading();
         xv = pose.getX();
         yv = pose.getY();
@@ -49,5 +49,21 @@ public class Robot implements Loggable {
 
     public Follower getFollower() {
         return follower;
+    }
+
+    public void snail() {
+        follower.setMaxPowerScaling(Setup.DriveSettings.SNAIL_SPEED);
+    }
+
+    public void normal() {
+        follower.setMaxPowerScaling(Setup.DriveSettings.NORMAL_SPEED);
+    }
+
+    public void auto() {
+        follower.setMaxPowerScaling(Setup.DriveSettings.AUTO_SCALING);
+    }
+
+    public void turbo() {
+        follower.setMaxPowerScaling(Setup.DriveSettings.TURBO_SPEED);
     }
 }
