+++ LearnBot/src/main/java/org/firstinspires/ftc/learnbot/controls/SafetyTestController.java
@@ -19,9 +19,6 @@ public class SafetyTestController {
         gamepad = g;
 
         AssignNamedControllerButton();
-        if (Setup.Connected.SAFETYSUBSYSTEM) {
-            bindDriveControls();
-        }
     }
 
     private void AssignNamedControllerButton() {
