+++ LearnBot/src/main/java/org/firstinspires/ftc/learnbot/opmodes/DriveBaseValidation.java
@@ -1,6 +1,7 @@
 package org.firstinspires.ftc.learnbot.opmodes;
 
 import com.bylazar.configurables.annotations.Configurable;
+import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 import com.qualcomm.robotcore.hardware.DcMotorEx;
@@ -10,6 +11,7 @@ import org.firstinspires.ftc.learnbot.Setup;
 import org.firstinspires.ftc.learnbot.subsystems.TestSubsystem;
 import org.firstinspires.ftc.robotcore.external.navigation.AngleUnit;
 
+@Disabled
 @Configurable
 @TeleOp(name = "Drivebase Testbed")
 public class DriveBaseValidation extends LinearOpMode {
