+++ LearnBot/src/main/java/org/firstinspires/ftc/learnbot/opmodes/JustDriveTele.java
@@ -1,0 +1,54 @@
+package org.firstinspires.ftc.learnbot.opmodes;
+
+import static org.firstinspires.ftc.learnbot.Setup.HardwareNames.AprilTag_Pipeline;
+
+import com.acmerobotics.dashboard.FtcDashboard;
+import com.acmerobotics.dashboard.telemetry.MultipleTelemetry;
+import com.bylazar.configurables.annotations.Configurable;
+import com.bylazar.gamepad.GamepadManager;
+import com.bylazar.gamepad.PanelsGamepad;
+import com.bylazar.telemetry.PanelsTelemetry;
+import com.bylazar.telemetry.TelemetryManager;
+import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
+import com.technototes.library.logger.Loggable;
+import com.technototes.library.structure.CommandOpMode;
+import com.technototes.library.util.Alliance;
+import org.firstinspires.ftc.learnbot.Hardware;
+import org.firstinspires.ftc.learnbot.Robot;
+import org.firstinspires.ftc.learnbot.controls.DriverController;
+import org.firstinspires.ftc.learnbot.helpers.StartingPosition;
+
+@Configurable
+@SuppressWarnings("unused")
+@TeleOp(name = "Just Drive")
+public class JustDriveTele extends CommandOpMode implements Loggable {
+
+    public Robot robot;
+    public DriverController controls;
+    public Hardware hardware;
+    // For Panels controller widget the two lines below
+    private final GamepadManager driverManager = PanelsGamepad.INSTANCE.getFirstManager();
+    private final TelemetryManager panelsTelemetry = PanelsTelemetry.INSTANCE.getTelemetry();
+
+    @Override
+    public void uponInit() {
+        telemetry = new MultipleTelemetry(
+            telemetry,
+            FtcDashboard.getInstance().getTelemetry(),
+            PanelsTelemetry.INSTANCE.getFtcTelemetry()
+        );
+        hardware = new Hardware(hardwareMap);
+        robot = new Robot(hardware, Alliance.NONE, StartingPosition.Unspecified);
+        controls = new DriverController(driverGamepad, robot);
+    }
+
+    @Override
+    public void uponStart() {
+        robot.atStart();
+    }
+
+    @Override
+    public void runLoop() {
+        telemetry.update();
+    }
+}
