+++ LearnBot/src/main/java/org/firstinspires/ftc/learnbot/opmodes/TestPedroAuto.java
@@ -2,7 +2,6 @@ package org.firstinspires.ftc.learnbot.opmodes;
 
 import com.acmerobotics.dashboard.FtcDashboard;
 import com.acmerobotics.dashboard.telemetry.MultipleTelemetry;
-import com.pedropathing.geometry.Pose;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.technototes.library.command.CommandScheduler;
 import com.technototes.library.command.SequentialCommandGroup;
@@ -11,13 +10,13 @@ import com.technototes.library.structure.CommandOpMode;
 import com.technototes.library.util.Alliance;
 import org.firstinspires.ftc.learnbot.Hardware;
 import org.firstinspires.ftc.learnbot.Robot;
+import org.firstinspires.ftc.learnbot.TestPaths;
 import org.firstinspires.ftc.learnbot.commands.PedroPathCommand;
-import org.firstinspires.ftc.learnbot.commands.auto.TestPaths;
 import org.firstinspires.ftc.learnbot.controls.DriverController;
 import org.firstinspires.ftc.learnbot.helpers.HeadingHelper;
 import org.firstinspires.ftc.learnbot.helpers.StartingPosition;
 
-@Autonomous(name = "TestPedro", preselectTeleOp = "OneDriver")
+@Autonomous(name = "Test Follow Speeds", preselectTeleOp = "Just Drive")
 @SuppressWarnings("unused")
 public class TestPedroAuto extends CommandOpMode {
 
@@ -31,21 +30,20 @@ public class TestPedroAuto extends CommandOpMode {
         hardware = new Hardware(hardwareMap);
         robot = new Robot(hardware, Alliance.RED, StartingPosition.Net);
         TestPaths p = new TestPaths(robot.follower);
-        CommandScheduler.scheduleInit(
-            new SequentialCommandGroup(
-                () -> robot.follower.setStartingPose(p.getStart()),
-                robot::updatePose
-            )
+        CommandScheduler.scheduleOnceForState(
+            () -> robot.follower.setStartingPose(p.getStart()),
+            OpModeState.INIT
         );
-        CommandScheduler.scheduleForState(robot::updatePose, OpModeState.RUN);
+        CommandScheduler.scheduleForState(robot::updatePose, OpModeState.INIT, OpModeState.RUN);
         CommandScheduler.scheduleForState(
             new SequentialCommandGroup(
+                robot::normal,
                 new PedroPathCommand(robot.follower, p.Path1),
-                new WaitCommand(0.5),
+                robot::snail,
                 new PedroPathCommand(robot.follower, p.Path2),
-                new WaitCommand(1),
+                robot::turbo,
                 new PedroPathCommand(robot.follower, p.Path3),
-                new WaitCommand(2),
+                robot::auto,
                 new PedroPathCommand(robot.follower, p.Path4),
                 CommandScheduler::terminateOpMode
             ),
