--- LearnBot/src/main/java/org/firstinspires/ftc/learnbot/subsystems/OdoAccuracy.md
@@ -1,34 +1,0 @@
-# Odo Accuracy
-
-Checking the accuracy of our odo pods when rotating
-
-5 rotations = 5 circumferences = $10 \pi r$
-
-421081 -> Radius is calculated as 13403 ticks 103964 -> Radius is calculated as
-3309 ticks
-
-TicksPerRev = 8192 WheelRadius = 17.5 mm
-
-Observed Radius is 1.63611 encoder revolutions Observed Radius is .404 encoder
-revolutions
-
-A revolution is $35\pi$ mm
-
-Observed Radius is 179.9 mm Observed Radius is 44.4 mm
-
-For distance, recall that $x^2 + y^2 = r^2$ thus:
-$x^2 + y^2 = 13403^2 = 179640409$ $x^2 + y^2 = 3309^2 =  10949481$
-
-Metal chassis is 31.2cm wide The drive base axles are 33.6cm apart
-
-Measured "outer" pod is 17.8cm x 5.2cm from center Measured "inner" pod is 7.6cm
-x 4.9cm from center
-
-#### Turning "error" (slippage on my floor):
-
-Outer pod -> 0.55cm radius, or about 3% error Inner pod -> 4.6cm radius, or
-about 104% error
-
-Graph of some of this (how I calculated error, in particular)
-
-https://www.desmos.com/calculator/laonl0zuwr
